Use a single socket for relaying
